Add activity with ViewPager2 and move event log classes to a separate package.
