+++ android/BOINC/app/src/main/java/edu/berkeley/boinc/BOINCActivity.kt
@@ -45,6 +45,7 @@ import edu.berkeley.boinc.client.ClientStatus
 import edu.berkeley.boinc.client.IMonitor
 import edu.berkeley.boinc.client.Monitor
 import edu.berkeley.boinc.databinding.MainBinding
+import edu.berkeley.boinc.ui.eventlog.EventLogActivity
 import edu.berkeley.boinc.utils.Logging
 import edu.berkeley.boinc.utils.RUN_MODE_AUTO
 import edu.berkeley.boinc.utils.RUN_MODE_NEVER
