+++ android/BOINC/app/src/main/java/edu/berkeley/boinc/ui/eventlog/EventLogActivity2.kt
@@ -1,0 +1,50 @@
+/*
+ * This file is part of BOINC.
+ * http://boinc.berkeley.edu
+ * Copyright (C) 2020 University of California
+ *
+ * BOINC is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License
+ * as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later version.
+ *
+ * BOINC is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with BOINC.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package edu.berkeley.boinc.ui.eventlog
+
+import android.os.Bundle
+import android.view.Menu
+import androidx.appcompat.app.AppCompatActivity
+import com.google.android.material.tabs.TabLayoutMediator
+import edu.berkeley.boinc.R
+import edu.berkeley.boinc.databinding.ActivityEventLog2Binding
+
+class EventLogActivity2 : AppCompatActivity() {
+    private lateinit var binding: ActivityEventLog2Binding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityEventLog2Binding.inflate(layoutInflater)
+        setContentView(binding.root)
+        val eventLogPagerAdapter = EventLogPagerAdapter(this)
+        binding.viewPager.adapter = eventLogPagerAdapter
+        TabLayoutMediator(binding.tabs, binding.viewPager) { tab, position ->
+            tab.text = if (position == 0)
+                getString(R.string.eventlog_client_header)
+            else
+                getString(R.string.eventlog_gui_header)
+        }.attach()
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.eventlog_menu, menu)
+        return true
+    }
+}
