+++ android/BOINC/app/src/main/java/edu/berkeley/boinc/ui/eventlog/EventLogGuiFragment.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with BOINC.  If not, see <http://www.gnu.org/licenses/>.
  */
-package edu.berkeley.boinc;
+package edu.berkeley.boinc.ui.eventlog;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -32,6 +32,7 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
+import edu.berkeley.boinc.R;
 import edu.berkeley.boinc.adapter.GuiLogRecyclerViewAdapter;
 import edu.berkeley.boinc.databinding.EventLogGuiLayoutBinding;
 import edu.berkeley.boinc.utils.BOINCUtils;
